Fix LConsole being bypassed to print exceptions on Frontend
LICC itself printed exceptions in multiple ways, without always considering the frontend type. Some need the stacktrace others hide it.

Bypassing LConsole, does prevent the calling of callback events for exceptions, which I added in this commit.

Using LineWriter to print generic exceptions, is not quite possible, as segments can only contain strings.
Thus Frontends are in charge of formatting exceptions in a style that suits them.

The new LConsole method starts with Print and not Write, as you are not in control - just like with a Printer :P