Change how the primary Frontend is managed
There is one primary frontend, associated with LConsole.
CommandConsole sets that frontend in LConsole, implying that there also is only one primary CommandConsole.
LICC tried to make the frontend swappable, but the attempt does not seem complete and reliable.

This commit removes the responsibility of the primary frontend from CommandConsole, one could now have multiple CommandConsoles (ignoring the static instance).

Instead the primary console is set and initialized in new FrontendManager. Which initializes the frontend as well as its history.
A frontend must have a history, else it potentially might break.

This change allows the primary frontend to be initialized and used, before the command framework is created. Which allows the usage of LConsole, in that time.

This solution is not perfect, but imo it was not perfect before either.
SwitchFrontend has been removed, as it was not complete.